// frontend/src/app/models/provider.model.ts
export interface Provider {
  firstName: string;
  lastName: string;
  ethnicIdentity: string;
  genderIdentity: string;
  availableCapacity: number;
  languages: string[];
  location: string;
  bio: string;
  sexualOrientation: string;
  religiousBackground: string;
  treatmentModalities: string[];
  specializationAreas: string[];
}